LABORATÓRIO 04/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
pyyaml==6.0.2
